+++ Assets/JFrameworkNet/Editor/Process/ServerRpcProcess.cs
@@ -41,15 +41,16 @@ namespace JFramework.Editor
             {
                 return null;
             }
-
+            
             worker.Emit(OpCodes.Ldarg_0);
             worker.Emit(OpCodes.Ldstr, md.FullName);
-            worker.Emit(OpCodes.Ldc_I4, NetworkEvent.GetHashByName(md.FullName));
+            worker.Emit(OpCodes.Ldc_I4, (int)NetworkEvent.GetHashByName(md.FullName));
             worker.Emit(OpCodes.Ldloc_0);
             worker.Emit(OpCodes.Ldc_I4, commandAttr.GetField("channel", 1));
             worker.Emit(OpCodes.Call, processor.sendServerRpcInternal);
             NetworkBehaviourProcess.WriteReturnWriter(worker, processor);
             worker.Emit(OpCodes.Ret);
+           
             return rpc;
         }
